fix : 패신저 예약 화면

+++ app/src/main/java/com/mate/carpool/data/repository/PassengerRepositoryImpl.kt
@@ -1,7 +1,5 @@
 package com.mate.carpool.data.repository
 
-import android.util.Log
-import android.widget.Toast
 import com.mate.carpool.data.model.DTO.TicketDeleteMemberRequestDTO
 import com.mate.carpool.data.model.DTO.TicketNewMemberRequestDTO
 import com.mate.carpool.data.model.response.ApiResponse
@@ -9,17 +7,16 @@ import com.mate.carpool.data.model.response.ResponseMessage
 import com.mate.carpool.data.service.APIService
 import com.mate.carpool.ui.utils.HandleFlowUtils.handleFlowApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class PassengerRepositoryImpl @Inject constructor(private val apiService: APIService) : PassengerRepository {
 
-    override fun addNewPassengerToTicket(id:Int): Flow<ApiResponse<ResponseMessage>> = handleFlowApi{
+    override fun addNewPassengerToTicket(id:Long): Flow<ApiResponse<ResponseMessage>> = handleFlowApi{
         apiService.postPassengerNew(TicketNewMemberRequestDTO(id))
     }
 
-    override fun deletePassengerToTicket(ticketId:Int,passengerId:Int): Flow<String> = handleFlowApi {
+    override fun deletePassengerToTicket(ticketId:Long,passengerId:Long): Flow<String> = handleFlowApi {
         apiService.deletePassenger(TicketDeleteMemberRequestDTO(passengerId,ticketId,0))
     }.map {
         when(it){
